+++ mapbox_maps_flutter_mobile/ios/mapbox_maps_flutter_mobile/Package.swift
@@ -3,7 +3,7 @@
 
 import PackageDescription
 
-let mapboxMapsVersion: Version = "11.27.0-SNAPSHOT-07-11--02-03.git-f207f96"
+let mapboxMapsVersion: Version = "11.28.0-SNAPSHOT-07-16--02-03.git-f78028c"
 
 let mapboxMapsPackage = mapboxMapsVersion.description.contains("SNAPSHOT")
     ? "mapbox-maps-ios-binary"