+++ mapbox_maps_flutter_mobile/ios/mapbox_maps_flutter_mobile/Sources/mapbox_maps_flutter_mobile/Classes/CompassController.swift
@@ -6,20 +6,15 @@ final class CompassController: CompassSettingsInterface {
 
     func updateSettings(settings: CompassSettings) throws {
         var compass = ornaments.options.compass
-        switch settings.position {
-        case .bOTTOMLEFT:
-            compass.position = .bottomLeading
-            compass.margins = CGPoint(x: settings.marginLeft ?? 0, y: settings.marginBottom ?? 0)
-        case .bOTTOMRIGHT:
-            compass.position = .bottomTrailing
-            compass.margins = CGPoint(x: settings.marginRight ?? 0, y: settings.marginBottom ?? 0)
-        case .tOPLEFT:
-            compass.position = .topLeading
-            compass.margins = CGPoint(x: settings.marginLeft ?? 0, y: settings.marginTop ?? 0)
-        case .tOPRIGHT, .none:
-            compass.position = .topTrailing
-            compass.margins = CGPoint(x: settings.marginRight ?? 0, y: settings.marginTop ?? 0)
+        if let position = settings.position {
+            compass.position = toNativeOrnamentPosition(position)
         }
+        compass.margins = margins.apply(
+            marginLeft: settings.marginLeft,
+            marginTop: settings.marginTop,
+            marginRight: settings.marginRight,
+            marginBottom: settings.marginBottom,
+            for: compass.position)
 
         if let data = settings.image?.data {
             compass.image = UIImage(data: data, scale: UIScreen.main.scale)
@@ -68,12 +63,12 @@ final class CompassController: CompassSettingsInterface {
         }
 
         return CompassSettings(
-            enabled: true,
+            enabled: visibility,
             position: position,
-            marginLeft: options.margins.x,
-            marginTop: options.margins.y,
-            marginRight: options.margins.x,
-            marginBottom: options.margins.y,
+            marginLeft: margins.left,
+            marginTop: margins.top,
+            marginRight: margins.right,
+            marginBottom: margins.bottom,
             opacity: 1,
             rotation: 0,
             visibility: visibility,
@@ -83,23 +78,12 @@ final class CompassController: CompassSettingsInterface {
         )
     }
 
-    func getFLT_SETTINGSOrnamentPosition(position: MapboxMaps.OrnamentPosition) -> OrnamentPosition {
-        switch position {
-        case .bottomLeading:
-            return .bOTTOMLEFT
-        case  .bottomTrailing:
-            return .bOTTOMRIGHT
-        case .topLeading:
-            return .tOPLEFT
-        default:
-            return.tOPRIGHT
-        }
-    }
-
     private var ornaments: OrnamentsManager
     private var cancelable: Cancelable?
+    private var margins: OrnamentMargins
 
     init(withMapView mapView: MapView) {
         self.ornaments = mapView.ornaments
+        self.margins = OrnamentMargins(seedingFrom: mapView.ornaments.options.compass.margins)
     }
 }